+++ server/src/controllers/auth.controller.js
@@ -263,6 +263,15 @@ export const adminLogin = async (req, res) => {
   }
 };
 
+export const adminLogout = (req, res) => {
+  try {
+    res.clearCookie("token");
+    res.status(200).json({ message: "Logged out successfully" });
+  } catch (error) {
+    res.status(500).json({ message: "internal server error" });
+  }
+};
+
 export const logout = (_req, res) => {
   try {
     res.clearCookie("token");