+++ server/src/routes/auth.route.js
@@ -1,6 +1,6 @@
 import express from "express";
 import rateLimit from 'express-rate-limit';
-import { signup, sendOTP2, login, adminLogin, logout, deleteAcc, updateProfile, updatePassword, checkAuth, verifyOTP, verifyOTP2 } from "../controllers/auth.controller.js";
+import { signup, sendOTP2, login, adminLogin, adminLogout, logout, deleteAcc, updateProfile, updatePassword, checkAuth, verifyOTP, verifyOTP2 } from "../controllers/auth.controller.js";
 import { authenticate } from "../middleware/auth.middleware.js";
 
 const router = express.Router();
@@ -22,6 +22,8 @@ router.post("/verify-otp2", verifyOTP2);
 router.post("/login", login);
 
 router.post("/adminLogin", adminLogin);
+
+router.post("/adminLogout", adminLogout);
 
 router.post("/logout", logout);