+++ History.md
@@ -1,4 +1,12 @@
-# 1.1.0 / 2010-02-08
+# 1.2.0 / 2025-03-05
+
+*Added join room feature
+*Update admin dashboard has been modefied
+*Fixed creating room functnality
+*Fixed Admin login bug 
+*Fixed minor bugs for improved stability
+
+# 1.1.0 / 2025-02-08
 
 *Added admin route security with custom admin authentication middleware  
 *Added notification feature for new messages